Add plugin metadata into the controller
We call it a generic plugin mechanism as there maybe other WASM related
functionality besides transforms that we want to give users.

This follows the standard backend,table,frontend pattern of handling
controller log logic. I expect the rest of the system to only need
access to the frontend, and we can choose to proxy the table requests
through it (the existing topic_table seems to be exposed directly in a
few places and I think it's nice to not worry about needing the frontend
and table and the differences in functionality between them personally).

I diverged a bit from the norm in terms of taking pointers instead of
sharded service references. I think it's a bit easier to understand
over overall (you only get a reference to your core's data) and it avoids
a bunch of clang tidy warnings about references as data members.

Would love to get feedback on this pattern.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>