Initial transform SDK
This defines the initial SDK that we expect users to develop with to
build data transforms.

The ABI is specified by a binary encoding of individual records that is
the same as the kafka wire protocol. However, unlike the kafka protocol,
we don't serialize whole batches to reduce the amount of memory that a
transform needs, we don't expect users to do anything different with a
single record vs a batch of records, and the FFI overhead is very low (
very little difference than a "normal" function call in C++), so the
tradeoff here should check out.

You can build these using "normal" golang, but that will stub out the
ABI calls that the redpanda host will provide, otherwise it's expected
that functional versions of this will be built using tinygo.

This is licensed under Apache 2 because we expect people to link this
into the transforms they build.

There are also a variety of transforms that are useful for testing that
can be found in the testdata directory. These are used in the tests
within redpanda proper.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>