Add the ability to recieve notifications for partition writes
In WASM transforms, we're going to want to be able to know when a write
has been completed and we can go read it and perform the transform. This
patch adds the ability to do this without polling.

However, this notification is best effort only, there is no garentee
that this is going to successfully fire for every write. In transform
land we'll solve this by having the transform periodically poll the
partition on a very slow interval (10 seconds or something?) to catch
any missing notifications. However on partitions with non-trivial
traffic this notification will fire on writes that take place later.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>
(cherry picked from commit da7237e152d70e525536f1e166b91d9a49151b55)