Add plugin functionality to controller
This is the metadata storage needed for WASM transforms. This also
adds the appropriate validation to topics and transforms themselves.

We add a new batch type for these controller command for replicating
these in the controller log using the established pattern. The
plugin_backend writes the data into plugin_table, and all requests to
this data need to be proxied through to the plugin_frontend, which
handles reads and writes to plugin metadata.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>