ssx: Add work_queue for running sequential tasks
Managing state when multiple futures can be in flight on mutable
functions in a class can be a headache requiring locks or very careful
modification of data structures.

A simpler pattern is one where all the public methods of a class
immediately forward work to a queue that forces linearization. This
removes a lot of mental overhead with needing to lock or understand how
the class is used by all callers.

This approach, while simplifying can lead to performance bottlenecks
when there is any I/O. It's usually better to explicitly have those
tasks execute outside of the queue. It might be possible that we want
this work queue to have explicit support for "run this task outside the
queue and report it's result back on the queue" but for now we forgo
that optimization.

This queue style is a very nice primative for control plane work where
the complexity lies in managing state and isn't directly tied to
client traffic.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>