Add alternate record_header constructor
When building batches, it's useful to just use the sizes of the iobuf,
which is true 99% of the time. The only time it's different is to
distinguish between null and empty iobuf (null iobuf has a size of -1),
and this constructor is opinionated to only support empty iobuf.

It may make sense to change the structures here to be
std::optional<iobuf> instead of being so tied to the kafka wire format.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>