model/transform: add a holder for user transformed data
The new ABI is going to decode timestamps, offset, etc in the host, then
the only binary data passed back and forth is going to be the payload
(the tuple of key, value, headers in a record).

This is a holder for the transform output, that forces validation of the
data to be in the correct format.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>