container: introduce chunked_vector
Introduce a fragmented_vector that doubles in size for the first chunk,
then after that allocates full chunks. The chunk size is always fixed
to our maximum recommended allocation size (128KB).

This implmentation relies on the underlying vector using a 2x (elements)
growth rate for it's underlying capacity, as to ensure we never
overshoot our max alloc size.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>