wasm: pass back output topic name for transform
In order to support writing to multiple output topics we need to know
which output topic to send the emitted record. Allow passing that data
around (right now it's always nullopt because we need new ABI methods to
support adding the topic).

We make this a view so we don't have to copy bytes out of the VM.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>