transform/processor: support specifying topic name
The wasm transform callback can now return topic names, so we need to be
able to handle an explicitly specified name. In order to do that keep,
sinks in a map and do lookup.

Also due to this change breakup batching to happen at the sink instead
of during the transform. This will allow transforms to do better
batching, but we need to be careful not to over batch. Future work will
plumb in max batch sizes so we will not overbatch.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>