transform/processor/test: don't enforce batching
It should not be valid to require the same batch strategy (IE all input
records in a single batch will be in the same output batch). And indeed
this will not always be the case later on, so instead make the
infrastructure for tests to work on a per record basis instead of a per
batch one.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>