transform/transfer_queue: remove semaphore
`ss::semaphore::wait` with an abort_source has some
stack-use-after-return issue that we haven't yet been able to track
down. Instead manually implement the semaphore pattern using a condition
variable. Since this is only used in a SPSC queue context, we can reuse
the existing condition_variable for this. If this was used with multiple
producers or multiple consumers it would be prone to races between when
waiters were unblocked and when `_used_memory` was mutated.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>