serde/parquet: record shredding algorithm implementation
This is the implementation of the record shredding algorithm from
the dremel paper, with the (quite good) example from the paper working.

The rules are fairly straightforward once you wrap your head around
them, but best explaination that I've found for this is found here:
https://github.com/julienledem/redelm/wiki/The-striping-and-assembly-algorithms-from-the-Dremel-paper#repetition-and-definition-levels

Primarily for the repetition level, we're recording when we repeat a
value and at what depth we're repeating it.