parquet: remove map types
Proper maps and lists are not encoded in parquet, but are a logical
translation using logical types. See the link below, but this needs
to be encoded semantically at the conversion process between iceberg
and parquet - trying to implement iceberg semantics in the parquet
layer will complicate the algorithms here and deviate from existing
implementations/the spec. Instead we will do this logical type mapping
at conversion time from iceberg -> parquet.

https://github.com/apache/parquet-format/blob/master/LogicalTypes.md