bazel/pbgen: more portable files
When including the header in a proto source file, use a bare import to
import the header in the current directory. When mangling imports we
need to support different directories. Also fix mangled imports by
passing in the sources descriptor set instead of the raw proto files in
the bzl file.