bytes: introduce ioarray
LSM trees work on blocks of SST files, and these blocks require random
access. Introduce a random access version of iobuf called ioarray.
ioarray has a fixed size, and contains a number of chunks that are
strictly a maximum of our max allocation size. It also supports aligned
allocation of each chunk for scatter/gather IO. This will be the key
data structure/bytes container used on the read path in the LSM tree
implementation, so it supports a few methods like reading out a string
from it or reading a 32 bit integer, as both are used in the LSM tree
implementation.