lsm/core/internal: introduce key type
In LevelDB, keys are encoded strings with the user's key and two
additional pieces of information: the sequence number and the "type"
of the key (value or tombstone). However, we when we encode a key we
store the trailing bytes that contain the sequence number and key type
in such a way that `std::memcmp` compares the in ordering we want for
keys, which is `user_key ASC, seqno DESC, type DESC`. This is different
than LevelDB, which decodes the trailing numeric value in order to do
the comparison. They also support custom sorting orders, which is
something we don't need and will just use `<=>` (which invokes
`std::memcmp` under the hood). This simplification means we don't have
to pass around a comparator all around the stack and we can `std::sort`
internal keys and it "just works".

LevelDB Links:
- https://github.com/google/leveldb/blob/ac691084fdc5546421a55b25e7653d450e5a25fb/db/dbformat.h#L49
- https://github.com/google/leveldb/blob/ac691084fdc5546421a55b25e7653d450e5a25fb/db/dbformat.cc#L47-L63