lsm/block: add bloom filter block
Bloom filter blocks are used for `get` requests to filter out SST files
that don't contain a key. It will be used during SST building to
generate a series of blocks every configurable interval.

LevelDB Links:
https://github.com/google/leveldb/blob/main/table/filter_block.cc
https://github.com/google/leveldb/blob/main/util/bloom.cc

Note we use `num_probes` instead of `k` as that name makes more sense to
me. This is also what RocksDB uses for it's name of this variable.