lsm/core: introduce lookup result
Introduce a tristate type to encode the result type for `get`
operations. We can't just use `optional<iobuf>` because we need to know
if we need to stop or keep going. LevelDB uses an error type of NotFound
for this, but that's actually really awkward to encode this way, and we
treat other errors as exceptions (corruption, io error, etc), so using a
result type here ends up with a really nice interface.