lsm/db: introduce version edit
A version within an LSM tree is a snapshot of the files within the LSM
tree structure. The LSM tree will track all versions currently being
used (as iterators can keep older versions around to provide snapshot
isolation). This edit data structure is used to accumulate changes (such
as ones from compaction, etc) which can then be applied to create a new
version within the database. This also introduces `file_meta_data` which
is the main metadata we track within the LSM tree for each file present
in the tree.

LevelDB Link:
* https://github.com/google/leveldb/blob/main/db/version_edit.h