lsm/core: add merging iterator
The merging iterator is what is used to merge iterators between all the
levels as well as the memtable, etc for the main database iterator. It
does not suppress duplicates, as we want to see all versions for some
internal iterators. Later there will be another (final) iterator
implementation that will suppress the duplicates.

LevelDB Link:
* https://github.com/google/leveldb/blob/main/table/merger.cc