lsm/db: introduce version set
A version set holds all the live versions of a database at any given
point in time. Each version itself is immutable and read-only.

The major difference here is that we don't use a log for the manifest
updates, but everytime we make a full snapshot of the metadata on disk.
It's a much simpler model and also compatible with object storage. In
the future we can use a log for metadata writes similar to LevelDB.
Also we define the storage format as a protobuf to make reading from
other tools simpler, and support a more flexible compatibility policy
than our normal serde format.

LevelDB Link:
* https://github.com/google/leveldb/blob/main/db/version_set.cc