lsm/db: draw the rest of the owl
This adds the core of the LSM tree database.
The core operations to be invoked externally is quite simple: for reads
we merge all the levels (memtables and SST files) and create an
iterator or do a `get` operation. All writes directly go to the
memtable. Everything else is scheduling, managing and performing
compactions of data.

Right now compaction is fairly simple. There is only a single compaction
ever going on at once all run in a single fiber. There are some basic
rate limiting and backpressure in place if compaction cannot keep up with
ingress traffic.

Overtime I expect the compaction logic to evolve as we support
parallel or subcompactions and independently flush memtables and compact
lower levels of the LSM tre.

Lastly, we introduce a primative benchmarking tool to prove out
performance and verify database intregrity. This should be advanced to
do more work concurrently.

LevelDB Links:
* https://github.com/google/leveldb/blob/main/db/db_impl.cc